# src/infraestructure/frameworks/express/docs/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FastFood API",
        "description": "API for FastFood Tech Challenge",
        "version": "1.0",
        "contact": {}
    },
    "servers": [],
    "paths": {
        "/products": {
            "post": {
                "operationId": "createProduct",
                "summary": "Create a new product",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateProductDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductPresenter"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "products"
                ]
            },
            "get": {
                "operationId": "getProducts",
                "summary": "Get products by category",
                "parameters": [
                    {
                        "name": "category",
                        "required": false,
                        "in": "query",
                        "description": "Filter products by category",
                        "schema": {}
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Products found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductPresenter"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid category",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "products"
                ]
            }
        },
        "/products/{id}": {
            "put": {
                "operationId": "updateProduct",
                "summary": "Update an existing product",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "ID of the product to update",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductPresenter"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "products"
                ]
            },
            "delete": {
                "operationId": "deleteProduct",
                "summary": "Delete a product",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "description": "ID of the product to delete",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted"
                    },
                    "404": {
                        "description": "Product not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "products"
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "CreateProductDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Name of the product",
                        "example": "Burguer Bacon"
                    },
                    "description": {
                        "type": "string",
                        "description": "Description of the product",
                        "example": "Contém 2 hamburgueres com queijo e bacon"
                    },
                    "category": {
                        "type": "string",
                        "description": "Category of the product",
                        "example": "Lanche",
                        "enum": [
                            "lanche",
                            "bebida",
                            "acompanhamento"
                        ]
                    },
                    "price": {
                        "type": "number",
                        "description": "Price of the product",
                        "example": 24.5
                    }
                },
                "required": [
                    "name",
                    "description",
                    "category",
                    "price"
                ]
            },
            "ProductPresenter": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "example": "1a442d43-fad4-47ad-91ce-485346ad5a05"
                    },
                    "name": {
                        "type": "string",
                        "example": "Burguer Bacon"
                    },
                    "description": {
                        "type": "string",
                        "example": "Contém 2 hamburgueres com queijo e bacon"
                    },
                    "category": {
                        "type": "string",
                        "example": "Lanche",
                        "enum": [
                            "lanche",
                            "acompanhamento",
                            "bebida"
                        ]
                    },
                    "price": {
                        "type": "string",
                        "example": "10.00"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "description",
                    "category",
                    "price"
                ]
            },
            "ErrorResponse": {
                "type": "object",
                "properties": {
                    "message": {
                        "type": "string",
                        "description": "Error message"
                    },
                    "statusCode": {
                        "type": "number",
                        "description": "HTTP status code"
                    }
                },
                "required": [
                    "message",
                    "error",
                    "statusCode"
                ]
            },
            "UpdateProductDto": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Name of the product",
                        "example": "Burguer Bacon"
                    },
                    "description": {
                        "type": "string",
                        "description": "Description of the product",
                        "example": "Contém 2 hamburgueres com queijo e bacon"
                    },
                    "category": {
                        "type": "string",
                        "description": "Category of the product",
                        "example": "Lanche",
                        "enum": [
                            "lanche",
                            "bebida",
                            "acompanhamento"
                        ]
                    },
                    "price": {
                        "type": "number",
                        "description": "Price of the product",
                        "example": 24.5
                    }
                },
                "required": [
                    "name",
                    "description",
                    "category",
                    "price"
                ]
            },
...
        }
    }
}